+++ tests/NCronJob.Tests/RuntimeJobRegistryTests.cs
@@ -7,6 +7,9 @@ namespace NCronJob.Tests;
 
 public class RuntimeJobRegistryTests : JobIntegrationBase
 {
+    private const string AtEveryMinute = "* * * * *";
+    private const string AtMinute2 = "2 * * * *";
+
     [Fact]
     public async Task DynamicallyAddedJobIsExecuted()
     {
@@ -15,7 +18,7 @@ public class RuntimeJobRegistryTests : JobIntegrationBase
         await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
         var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
 
-        registry.TryRegister(s => s.AddJob(async (ChannelWriter<object> writer) => await writer.WriteAsync(true), "* * * * *"), out _);
+        registry.TryRegister(s => s.AddJob(async (ChannelWriter<object> writer) => await writer.WriteAsync(true), AtEveryMinute), out _);
 
         FakeTimer.Advance(TimeSpan.FromMinutes(1));
         var jobFinished = await WaitForJobsOrTimeout(1);
@@ -30,8 +33,8 @@ public class RuntimeJobRegistryTests : JobIntegrationBase
         await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
         var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
 
-        registry.TryRegister(s => s.AddJob(async (ChannelWriter<object> writer) => await writer.WriteAsync(true), "* * * * *"));
-        registry.TryRegister(s => s.AddJob(async (ChannelWriter<object> writer) => await writer.WriteAsync(true), "* * * * *"));
+        registry.TryRegister(s => s.AddJob(async (ChannelWriter<object> writer) => await writer.WriteAsync(true), AtEveryMinute));
+        registry.TryRegister(s => s.AddJob(async (ChannelWriter<object> writer) => await writer.WriteAsync(true), AtEveryMinute));
 
         FakeTimer.Advance(TimeSpan.FromMinutes(1));
         var jobFinished = await WaitForJobsOrTimeout(2);
@@ -42,7 +45,7 @@ public class RuntimeJobRegistryTests : JobIntegrationBase
     public async Task CanRemoveJobByName()
     {
         ServiceCollection.AddNCronJob(
-            s => s.AddJob(async (ChannelWriter<object> writer) => await writer.WriteAsync(true), "* * * * *", jobName: "Job"));
+            s => s.AddJob(async (ChannelWriter<object> writer) => await writer.WriteAsync(true), AtEveryMinute, jobName: "Job"));
         var provider = CreateServiceProvider();
         await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
         var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
@@ -55,9 +58,25 @@ public class RuntimeJobRegistryTests : JobIntegrationBase
     }
 
     [Fact]
+    public async Task DoesNotCringeWhenRemovingNonExistingJobs()
+    {
+        ServiceCollection.AddNCronJob();
+
+        var provider = CreateServiceProvider();
+        await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
+        var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
+
+        var jobRegistry = provider.GetRequiredService<JobRegistry>();
+        Assert.Empty(jobRegistry.GetAllJobs());
+
+        registry.RemoveJob("Nope");
+        registry.RemoveJob<SimpleJob>();
+    }
+
+    [Fact]
     public async Task CanRemoveByJobType()
     {
-        ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p.WithCronExpression("* * * * *")));
+        ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p.WithCronExpression(AtEveryMinute)));
         var provider = CreateServiceProvider();
         await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
         var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
@@ -70,6 +89,24 @@ public class RuntimeJobRegistryTests : JobIntegrationBase
     }
 
     [Fact]
+    public async Task RemovingByJobTypeAccountsForAllJobs()
+    {
+        ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p.WithCronExpression("1 * * * *")));
+        ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p.WithCronExpression(AtMinute2)));
+
+        var provider = CreateServiceProvider();
+        await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
+        var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
+
+        var jobRegistry = provider.GetRequiredService<JobRegistry>();
+        Assert.Equal(2, jobRegistry.FindAllJobDefinition(typeof(SimpleJob)).Count);
+
+        registry.RemoveJob<SimpleJob>();
+
+        Assert.Empty(jobRegistry.FindAllJobDefinition(typeof(SimpleJob)));
+    }
+
+    [Fact]
     public async Task CanUpdateScheduleOfAJob()
     {
         ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p.WithCronExpression("0 0 * * *").WithName("JobName")));
@@ -77,7 +114,12 @@ public class RuntimeJobRegistryTests : JobIntegrationBase
         await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
         var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
 
-        registry.UpdateSchedule("JobName", "* * * * *");
+        registry.UpdateSchedule("JobName", AtEveryMinute);
+
+        var jobRegistry = provider.GetRequiredService<JobRegistry>();
+        var jobDefinition = jobRegistry.GetAllJobs().Single();
+
+        Assert.Equal(AtEveryMinute, jobDefinition.UserDefinedCronExpression);
 
         FakeTimer.Advance(TimeSpan.FromMinutes(1));
         var jobFinished = await WaitForJobsOrTimeout(1);
@@ -92,7 +134,7 @@ public class RuntimeJobRegistryTests : JobIntegrationBase
         await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
         var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
 
-        Should.Throw<InvalidOperationException>(() => registry.UpdateSchedule("JobName", "* * * * *"));
+        Should.Throw<InvalidOperationException>(() => registry.UpdateSchedule("JobName", AtEveryMinute));
     }
 
     [Fact]
@@ -130,7 +172,7 @@ public class RuntimeJobRegistryTests : JobIntegrationBase
     {
         ServiceCollection.AddNCronJob(s =>
         {
-            s.AddJob<SimpleJob>(p => p.WithCronExpression("* * * * *").WithName("Job1"))
+            s.AddJob<SimpleJob>(p => p.WithCronExpression(AtEveryMinute).WithName("Job1"))
                 .ExecuteWhen(r => r.RunJob(() => { }, "Job2"));
         });
         var provider = CreateServiceProvider();
@@ -147,7 +189,7 @@ public class RuntimeJobRegistryTests : JobIntegrationBase
     public async Task UpdatingParameterHasImmediateEffect()
     {
         ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p
-            .WithCronExpression("* * * * *")
+            .WithCronExpression(AtEveryMinute)
             .WithParameter("foo")
             .WithName("JobName")));
         var provider = CreateServiceProvider();
@@ -170,7 +212,7 @@ public class RuntimeJobRegistryTests : JobIntegrationBase
             .WithName("JobName")));
         var provider = CreateServiceProvider();
         var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
-        registry.TryRegister(s => s.AddJob(() => { }, "* * * * *", jobName: "JobName2"), out _);
+        registry.TryRegister(s => s.AddJob(() => { }, AtEveryMinute, jobName: "JobName2"), out _);
 
         var allSchedules = registry.GetAllRecurringJobs();
 
@@ -179,7 +221,7 @@ public class RuntimeJobRegistryTests : JobIntegrationBase
                                         && s.CronExpression == "*/2 * * * *"
                                         && s.TimeZone == timeZone);
         allSchedules.ShouldContain(s => s.JobName == "JobName2"
-                                        && s.CronExpression == "* * * * *"
+                                        && s.CronExpression == AtEveryMinute
                                         && s.TimeZone == TimeZoneInfo.Utc);
     }
 
@@ -189,20 +231,20 @@ public class RuntimeJobRegistryTests : JobIntegrationBase
         ServiceCollection.AddNCronJob(p => p.AddJob<SimpleJob>());
         var provider = CreateServiceProvider();
         var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
-        registry.TryRegister(n => n.AddJob<SimpleJob>(p => p.WithCronExpression("* * * * *").WithName("JobName")));
+        registry.TryRegister(n => n.AddJob<SimpleJob>(p => p.WithCronExpression(AtEveryMinute).WithName("JobName")));
 
         var allSchedules = registry.GetAllRecurringJobs();
 
         allSchedules.Count.ShouldBe(1);
         allSchedules.ShouldContain(s => s.JobName == "JobName"
-                                        && s.CronExpression == "* * * * *"
+                                        && s.CronExpression == AtEveryMinute
                                         && s.TimeZone == TimeZoneInfo.Utc);
     }
 
     [Fact]
     public async Task ShouldDisableJob()
     {
-        ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p.WithCronExpression("* * * * *").WithName("JobName")));
+        ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p.WithCronExpression(AtEveryMinute).WithName("JobName")));
         var provider = CreateServiceProvider();
         await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
         var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
@@ -212,6 +254,49 @@ public class RuntimeJobRegistryTests : JobIntegrationBase
         FakeTimer.Advance(TimeSpan.FromMinutes(1));
         var jobFinished = await WaitForJobsOrTimeout(1, TimeSpan.FromMilliseconds(200));
         jobFinished.ShouldBeFalse();
+    }
+
+    [Fact]
+    public async Task DisablingAndEnablingByJobTypeAccountsForAllJobs()
+    {
+        ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>());
+        ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p.WithCronExpression(AtMinute2)));
+
+        var provider = CreateServiceProvider();
+        await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
+        var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
+
+        var jobRegistry = provider.GetRequiredService<JobRegistry>();
+        var jobs = jobRegistry.FindAllJobDefinition(typeof(SimpleJob));
+        Assert.Equal(2, jobs.Count);
+
+        Assert.All(jobs, j =>
+        {
+            if (j.CronExpression is null)
+            {
+                return;
+            }
+
+            Assert.NotEqual(RuntimeJobRegistry.TheThirtyFirstOfFebruary, j.CronExpression);
+        });
+
+        registry.DisableJob<SimpleJob>();
+
+        jobs = jobRegistry.FindAllJobDefinition(typeof(SimpleJob));
+        Assert.Equal(2, jobs.Count);
+
+        Assert.All(jobs, j =>
+        {
+            Assert.Equal(RuntimeJobRegistry.TheThirtyFirstOfFebruary, j.CronExpression);
+        });
+
+        registry.EnableJob<SimpleJob>();
+
+        jobs = jobRegistry.FindAllJobDefinition(typeof(SimpleJob));
+        Assert.Equal(2, jobs.Count);
+
+        Assert.Single(jobs, j => j.CronExpression is null);
+        Assert.Single(jobs, j => j.CronExpression is not null && j.CronExpression.ToString() == AtMinute2);
     }
 
     [Fact]
@@ -228,7 +313,7 @@ public class RuntimeJobRegistryTests : JobIntegrationBase
     [Fact]
     public async Task ShouldEnableJob()
     {
-        ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p.WithCronExpression("* * * * *").WithName("JobName")));
+        ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p.WithCronExpression(AtEveryMinute).WithName("JobName")));
         var provider = CreateServiceProvider();
         await provider.GetRequiredService<IHostedService>().StartAsync(CancellationToken);
         var registry = provider.GetRequiredService<IRuntimeJobRegistry>();
@@ -244,7 +329,7 @@ public class RuntimeJobRegistryTests : JobIntegrationBase
     [Fact]
     public void ShouldThrowWhenDuplicateJobNamesDuringRegistration()
     {
-        var act = () => ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p.WithCronExpression("* * * * *").WithName("JobName")
+        var act = () => ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p.WithCronExpression(AtEveryMinute).WithName("JobName")
             .And.WithCronExpression("*/2 * * * *").WithName("JobName")));
 
         act.ShouldThrow<InvalidOperationException>();
@@ -253,12 +338,12 @@ public class RuntimeJobRegistryTests : JobIntegrationBase
     [Fact]
     public void ShouldThrowRuntimeExceptionWithDuplicateJob()
     {
-        ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p.WithCronExpression("* * * * *").WithName("JobName")));
+        ServiceCollection.AddNCronJob(s => s.AddJob<SimpleJob>(p => p.WithCronExpression(AtEveryMinute).WithName("JobName")));
         var runtimeJobRegistry = CreateServiceProvider().GetRequiredService<IRuntimeJobRegistry>();
 
         var successful = runtimeJobRegistry.TryRegister(s => s.AddJob(() =>
         {
-        }, "* * * * *", jobName: "JobName"), out var exception);
+        }, AtEveryMinute, jobName: "JobName"), out var exception);
 
         successful.ShouldBeFalse();
         exception.ShouldNotBeNull();
@@ -270,9 +355,9 @@ public class RuntimeJobRegistryTests : JobIntegrationBase
     {
         ServiceCollection.AddNCronJob();
         var runtimeJobRegistry = CreateServiceProvider().GetRequiredService<IRuntimeJobRegistry>();
-        runtimeJobRegistry.TryRegister(s => s.AddJob<SimpleJob>(p => p.WithCronExpression("* * * * *")));
+        runtimeJobRegistry.TryRegister(s => s.AddJob<SimpleJob>(p => p.WithCronExpression(AtEveryMinute)));
 
-        var successful = runtimeJobRegistry.TryRegister(s => s.AddJob<SimpleJob>(p => p.WithCronExpression("* * * * *")), out var exception);
+        var successful = runtimeJobRegistry.TryRegister(s => s.AddJob<SimpleJob>(p => p.WithCronExpression(AtEveryMinute)), out var exception);
 
         successful.ShouldBeFalse();
         exception.ShouldNotBeNull();
@@ -284,7 +369,7 @@ public class RuntimeJobRegistryTests : JobIntegrationBase
         ServiceCollection.AddNCronJob();
         var runtimeJobRegistry = CreateServiceProvider().GetRequiredService<IRuntimeJobRegistry>();
 
-        var successful = runtimeJobRegistry.TryRegister(s => s.AddJob<SimpleJob>(p => p.WithCronExpression("* * * * *")), out var exception);
+        var successful = runtimeJobRegistry.TryRegister(s => s.AddJob<SimpleJob>(p => p.WithCronExpression(AtEveryMinute)), out var exception);
 
         successful.ShouldBeTrue();
         exception.ShouldBeNull();